Add routes for testArray and itemDisplay
