+++ src/Containers/App.js
@@ -5,11 +5,8 @@ import {
   Switch,
   Route,
   Link,
-  useRouteMatch,
-  useParams,
 } from 'react-router-dom';
-import SearchBar from '../Components/SearchBar';
-import Item from '../Components/Item';
+import ItemList from './ItemList';
 
 export default function App() {
   return (
@@ -20,80 +17,19 @@ export default function App() {
             <Link to="/">Home</Link>
           </li>
           <li>
-            <Link to="/item">Item</Link>
-          </li>
-          <li>
-            <Link to="/topics">Topics</Link>
+            <Link to="/itemList">itemList</Link>
           </li>
         </ul>
 
         <Switch>
-          <Route path="/item">
-            <Item />
-          </Route>
-          <Route path="/topics">
-            <Topics />
+          <Route path="/itemList">
+            <ItemList />
           </Route>
           <Route path="/">
-            <Home />
+            <div>Home</div>
           </Route>
         </Switch>
       </div>
     </Router>
-  );
-}
-
-function Home() {
-  return (
-    <div>
-      <h2>Home</h2>
-      <SearchBar />
-    </div>
-
-  );
-}
-
-function Topics() {
-  const match = useRouteMatch();
-  const itemList = ['apple', 'pear', 'banana', 'orange'];
-  return (
-    <div>
-      <h2>Topics</h2>
-
-      <ul>
-        <li>
-          <Link to={`${match.url}/${1 + 1}`}>2</Link>
-        </li>
-        <li>
-          <Link to={`${match.url}/props-v-state`}>
-            Props v. State
-          </Link>
-        </li>
-      </ul>
-
-      {/* The Topics page has its own <Switch> with more routes
-          that build on the /topics URL path. You can think of the
-          2nd <Route> here as an "index" page for all topics, or
-          the page that is shown when no topic is selected */}
-      <Switch>
-        <Route path={`${match.path}/:topicId`}>
-          <Topic />
-        </Route>
-        <Route path={match.path}>
-          <h3>Please select a topic.</h3>
-        </Route>
-      </Switch>
-    </div>
-  );
-}
-
-function Topic() {
-  const { topicId } = useParams();
-  return (
-    <h3>
-      Requested topic ID:
-      {' '}
-      {topicId}
-    </h3>
   );
 }
