+++ src/reducers/root.js
@@ -1,7 +1,9 @@
 import { combineReducers } from 'redux';
 import filterReducer from './filter';
+import itemReducer from './items';
 
 const rootReducer = combineReducers({
+  items: itemReducer,
   category: filterReducer,
 });
